Software runstop support via controller and topics
Allows breakers to be disabled via software. Re-enabling
can be done via software, or by toggling the physical runstop.

Teleop control and which breakers are controlled are toggled by
flags. Listens for boolean msgs on topic /enable_software_runstop.
(e.g. for interfacing with rqt_fetch_dash)

Script can be run separately, but is integrated with the
robot drivers, and does not by default enable runstopping with
the controller.

Includes unit tests for statemachine-like behavior of software
runstop.