[fetch_binary_drivers] new [fetch_drivers] package
This is a new package providing the "fetch_drivers" which was previously
built internally and distributed as a binary only debian package.

We will officially announce when the upgrade from Indigo->Melodic is
ready.

18.04 and ROS Melodic on the Fetch Research Platforms is still being
tested.

The process isn't straight foward as there are calibration files
which you will want to back-up before the upgrade.

This `fetch_drivers` package also won't start like it use to because of
the change from `upstart` to `systemd`.
That requires `fetch_system_config` which will be open sourced soon.

This `fetch_binary_drivers` package is part of a plan to distribute
updates and upgrades to Fetch Research Platform customers faster in the future.

This relates to fetchrobotics/fetch_ros#63

1. An officiall announcement will be made when everything is ready.
2. We will update and document the process on:
    https://docs.fetchrobotics.com/
3. We will also announce via our mailing list, ros discourse and post on:
    https://opensource.fetchrobotics.com/