Redo logging of fetch systemd services
Because the version of systemd that ships with Ubuntu 18.04 does not
support appending to log files when using the StandardOutput/Error
directives, we instead redirect logs from journalctl to syslog, and use
rsyslog to handle creation of actual log files.

In conjunction with this:
- robot.log moved from /var/log/ros/ to /var/log
- We stopped putting the logs for roscore and ps3joy in a file. The logs
of the robot service are commonly used for troubleshooting/support,
whereas the other services' logs are not generally informative.
- removed logrotate rule referencing old sixad upstart job.